// SnD/src/types/InputTypes.ts
/* eslint-disable no-shadow */
enum InputTypes {
    TEXT = 'text',
    DATE = 'date',
    DATE_TIME = 'datetime-local',
    TIME = 'time',
    NUMBER = 'number',
    MOBILE_NUMBER = 'mob',
    UNIT = 'uint',
    INTEGER = 'int',
    POSITIVE_CURRENCY_VALUE = 'ucurrency',
    POSITIVE_DOUBLE_VALUE = 'udouble',
    CURRENCY_VALUE = 'currency',
    DOUBLE_VALUE = 'double',
    TEXT_ARE = 'textarea',
    CHECKBOX = 'checkbox',
    EMAIL = 'email',
    FILE = 'file',
    BOOL = 'bool',
    ARRAY = 'Array<any>',
    DROP_ZONE = 'drop-zone'
}

export default InputTypes;